Add support for golang-based buildpacks which are using go modules
Requires Go 1.11 when running it, that is to generate and package the
vendor folder to be later used by OBS.

In the OBS spec, when a buildpack contains a ```go.mod``` file, we
reconstruct the GOPATH with the modules fetched previously required
while building both ```buildpack-packager``` and the buildpack itself.
This is also because buildpack-packager is not vendored as part of the
buildpacks anymore, and the same vendored modules are used for both the
buildpack and buildpack-packager.

Fetches ```buildpack-packager``` main.go during the process - as ```go
mod vendor``` filters out non-modules.