Use base stack for checking the build_status
otherwise we expect all supported stacks to work which is not always the
case.

This can quite be optimized. Instead of iterating about our stacks we
should iterate about the upstream ones, cflinuxfs2 and cflinuxfs3.

In this case we would have two runs instead of three and there we can
match multiple repositories.

They way we run it at the moment we can only check one repo at a time.